move HTML and Text components into the Emoji directory
